+++ src/features/Signup/stories/ShortPasswordForm.tsx
@@ -2,7 +2,7 @@ import { expect } from '@storybook/jest';
 import { StoryObj } from '@storybook/react';
 import { userEvent, within } from '@storybook/testing-library';
 
-import Signup from '../Signup';
+import Signup from '../components/pages/Signup';
 
 type Story = StoryObj<typeof Signup>;
